requirements.txt
pandas==2.0.3
numpy==1.26.4
plotly==5.15.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
scikit-learn==1.3.2
boto3==1.34.40
python-dotenv==1.0.1